setOutputTree.h (line 21): float OriginalWeight;
Typo in OriginalWeight => originalWeight

+++ WVAanalysisRun2/interface/setOutputTree.h
@@ -18,7 +18,7 @@ class setOutputTree {
   int isBoostedSignal;
   float wSampleWeight;
   float genWeight;
-  double originalWeight;
+  float originalWeight;
   std::vector<double> AQGCweights;
   float totalEventWeight;
   float totalEventWeight_2;
